raster file inheritance
Nog issue met lock door gdal.Open
